+++ CHANGELOG.md
@@ -25,3 +25,4 @@
  - sort by field name
  - query and sort with single command
  - query between dates
+ - convert negative limit to high integer number